Move Configuration to api package
